Done apis

+++ src/main/java/com/kimho/book/service/impl/CommentServiceImpl.java
@@ -2,7 +2,10 @@ package com.kimho.book.service.impl;
 
 import com.kimho.book.converter.CommentToDto;
 import com.kimho.book.converter.DtoToComment;
+import com.kimho.book.exception.NotFoundException;
+import com.kimho.book.exception.UnauthorizedException;
 import com.kimho.book.model.dao.Comment;
+import com.kimho.book.model.dao.User;
 import com.kimho.book.model.dto.CommentDto;
 import com.kimho.book.model.dto.CommentUpdate;
 import com.kimho.book.repository.CommentRepository;
@@ -21,6 +24,8 @@ public class CommentServiceImpl implements BooksService<CommentDto, CommentUpdat
     private CommentToDto commentToDto;
     @Autowired
     private DtoToComment dtoToComment;
+    @Autowired
+    private UserServiceImpl userService;
 
     @Override
     public List<CommentDto> getAll() {
@@ -29,17 +34,26 @@ public class CommentServiceImpl implements BooksService<CommentDto, CommentUpdat
 
     @Override
     public CommentDto get(long id) {
+        verifyCommentIsExist(id);
         return commentToDto.convert(commentRepository.findById(id).get());
     }
 
+    private Comment findById(long id){
+        verifyCommentIsExist(id);
+        return commentRepository.findById(id).get();
+    }
     @Override
     public CommentDto post(CommentDto commentDto) {
+        commentDto.setUserId(userService.getMyUser().getId());
         return commentToDto.convert(commentRepository.save(dtoToComment.convert(commentDto)));
     }
 
     @Override
     public CommentDto put(CommentUpdate commentEdition, long id) {
-        Comment comment = commentRepository.findById(id).get();
+        Comment comment = findById(id);
+        if (!checkAuthorized(comment.getUser())){
+            throw new UnauthorizedException("Unauthorized");
+        }
         comment.setMessage(commentEdition.getMessage());
         comment.setUpdatedAt(new Date());
         return commentToDto.convert(comment);
@@ -47,6 +61,28 @@ public class CommentServiceImpl implements BooksService<CommentDto, CommentUpdat
 
     @Override
     public void delete(long id) {
+        verifyCommentIsExist(id);
+        if (!checkAuthorized(findById(id).getUser())){
+            throw new UnauthorizedException("Unauthorized");
+        }
         commentRepository.deleteById(id);
+    }
+
+    private boolean checkAuthorized(User userCreateComment) {
+        User myUser = userService.getMyUser();
+        if (myUser.getId() == userCreateComment.getId()) {
+            return true;
+        } else if (myUser.isSuperAdmin()) {
+            return true;
+        } else if (myUser.isAdmin() && userCreateComment.isUser()) {
+            return true;
+        }
+        return false;
+    }
+
+    private void verifyCommentIsExist(long id) {
+        if (!commentRepository.existsById(id)) {
+            throw new NotFoundException("Comment not found");
+        }
     }
 }
